# Make getindex and setindex! for DenseTensor a bit faster