This PR makes `getindex`, `setindex!` and `iterate` for `DenseTensor` a bit faster.

For example, timing the following code:
```julia
using ITensors

function sum(A)
  x = zero(eltype(A))
  for i in A
    x += i
  end
  return x
end

let
  n = 200
  T = Tensor(n,n)
  randn!(T)
  A = randn(n,n)

  println("getindex Array")
  @btime $A[4,4]

  println("getindex Tensor")
  @btime $T[4,4]

  println("setindex! Array")
  @btime $A[4,4] = 2.0

  println("setindex! Tensor")
  @btime $T[4,4] = 2.0

  println("iterate Array")
  @btime sum($A)

  println("iterate Tensor")
  @btime sum($T)

  nothing
end
```

On master:
```julia
julia> include("run.jl")
getindex Array
  1.448 ns (0 allocations: 0 bytes)
getindex Tensor
  2.578 ns (0 allocations: 0 bytes)
setindex! Array
  1.415 ns (0 allocations: 0 bytes)
setindex! Tensor
  1.839 ns (0 allocations: 0 bytes)
iterate Array
  36.467 μs (0 allocations: 0 bytes)
iterate Tensor
  57.448 μs (0 allocations: 0 bytes)
```
and on this branch:
```julia
julia> include("run.jl")
getindex Array
  1.382 ns (0 allocations: 0 bytes)
getindex Tensor
  1.418 ns (0 allocations: 0 bytes)
setindex! Array
  1.415 ns (0 allocations: 0 bytes)
setindex! Tensor
  1.386 ns (0 allocations: 0 bytes)
iterate Array
  36.477 μs (0 allocations: 0 bytes)
iterate Tensor
  36.490 μs (0 allocations: 0 bytes)
```

For `getindex` and `setindex!`, the problem was that two bounds checks were happening, so I turned off one of them (first the CartesianIndex was being converted to a LinearIndex, which did one bounds check, then the storage was indexed with the LinearIndex, which did another). For `iterate`, the improvement was to make sure it was using Julia's `iterate` function for the `Vector` data, instead of a generic iteration function.

This speeds up DMRG a little bit. For example for the 1D spin-1 Heisenberg model with 100 sites and no QNs, on master:
```julia
After sweep 1 energy=-137.340778003834 maxlinkdim=9 time=0.150
After sweep 2 energy=-138.934529420032 maxlinkdim=20 time=0.410
After sweep 3 energy=-138.940074388108 maxlinkdim=95 time=1.493
After sweep 4 energy=-138.940086000816 maxlinkdim=100 time=3.273
After sweep 5 energy=-138.940086076384 maxlinkdim=96 time=3.776
```
and on this branch:
```julia
After sweep 1 energy=-137.663060169546 maxlinkdim=9 time=0.141
After sweep 2 energy=-138.933857520027 maxlinkdim=20 time=0.398
After sweep 3 energy=-138.940070648285 maxlinkdim=92 time=1.284
After sweep 4 energy=-138.940085932745 maxlinkdim=100 time=2.900
After sweep 5 energy=-138.940086074873 maxlinkdim=95 time=3.183
```